Add product searching functionality
Allow the user to enter a search query in the search box.

Aggregated the different states that are related to product fetching
to a single object, 'productState', since sometimes several values need
to change simultaneusly:
+ when search is initiated, the current product list needs to be cleared
+ when a tag filter is requested, the current product list needs to be
cleared

Since:
+ products need to be fetched only after certain kinds of
updates to productState
+ calling 'getProducts' inside a setState function does not take into
account the last state change (that is, the 'getProducts' call runs
with an outdated state)
I solved this issue by introducing another state variable,
'effectTriggerId', which is a dependency for the 'getProducts' side
effect. It consists of an ID that increments each time an update is
requested through the 'triggerProductUpdateAfterStateUpdate' function.

There may be a less convoluted way to achieve this.