Fix searchbox responsivity and superflous fetch requests to the API
Removed searchbox input validation since it was making input unresponsive.

Implemented a cache for the products that are fetched without a search
query, and restore the product list from the cache when the search box
is cleared to avoid having to fetch the same items from the server
time and time again if the user keeps writing and deleting queries.

The cache could be expanded to hold the result of all the queries.

Restructured the state of the product list to separate states, so that
side effects could be assigned independently to each one. This cleans
up the dummy state that was only used to signal that a new fetch should
happen, overall making the code logic much more easier to follow.